The SDL Renderer Beast (i fucking hate it here)
Ahh, one of the fun parts of SDL3 is trying to craft arguments that work, due to how much everything is different from SDL2.

Created a renderer instance successfully and am slowly implementing the game logic so that my code doesn't break. There's a lot of ways in which a piece of code can completely fuck over the rest of the project, because good practices in SDL3 is a hard-ass journey