App V0.2
One of the most important parts of handling this code is to get the functions to work with each other. The game logic is sound, but how the functions talk to each other is often less than idea. The biggest problem I was facing was a borked init function that is now fixed. Instead of error checking as means to start, we explicitly call for the window to be created in init, and then only error check this in main, where it actually matters.

Next up will be incorporating event handling and the renderer. However, I am slowly changing things so they don't become broken again