Rendering successful - update to x printing required
This is a major feat for the project, everything being properly rendered. I'd like to thank the Academy

Now, we have a game of Tic Tac Toe playable, despite some of the issues with the x not rendering how it should (most likely a mistake I made when reincorporating these functions after failing to get results).

This is a big move, and a big deal for what I'm working on