Important Game Screens
While I am working to implement the final features for my first game, I need a start, gameplay and an endscreen option.

I may end up doing a leaderboard that keeps track of who is winning in the game.

There are a few things left to implement:
- Audio
- Video (the meme)

This is all the frontend features I need before making it a game. For this game, the project scope is incredibly simple so I don't sit there trying to construct a 3D FPS game and abandon it because it is too difficult. It also is enabling me to have some prebuilt code so that for the next project, I can easily jump into implementation.

The goal is to have JUCE enabled to do audio event handling. Since it is a C++ library that shouldn't interfere with SDL, it'll be perfect