Meme Ending - Video/Audio Integration
Video and audio integration works well EXCEPT in the case of text overlaying.

I will modify this to do what is needed later. It is the bane of the project's existence, and is what causes the program to segfault if used.

There aren't too many good methods for doing this, so I'll have to carefully come up with something that works