Audio Event Handling V2
In this, we can explicitly call to the sound FX each time it's played. This follows some similar logic to my JavaScript Speed Pong game, because this was an issue from a previous project involving sound effect initialization.

It would work better if I did it restart style (in terms of a better modular approach