Failed App Build Attempt
Since this is so fucking complicated, and I've largely needed to rely upon LLMs to figure this part out, I'm going to learn the process of building a Hello World version of this process. The biggest problem is text, because why it fails to initialize will be the big factor.

Packaging an app is not simple, and hence, I will be slowly unraveling working with libraries and properly making an app build that is fucking compliant to SDL3.

It's not just that we are dealing with SDL as a source, but a multitude of different libraries, so hence, the slow process of unraveling where I am getting issues with adding SDL3_ttf. For some reason, this is a problem child.