Finalized Main
This version makes sure we only destroy text as an object upon destruction of the application. This prevents a rendering during each frame problem that was persistent. This allows us to cleanup the font only when it is called to be cleaned up.

Also did some nice formatting to make things look better