Makefile
This OFFICIALLY handles bundling the application AND adding an icon. This was such a problem getting this to work, and that is because photos are treated for icons very weirdly.

I also added some update commands so that you could refresh the icon's existence.

This game is complete, and all that there is left is to zip this into its own folder and upload to Itch.io. This project took so many commits, but it is finally complete